requirements.txt
flask
requests
python-dotenv
plotly
spotipy
numpy
python-dotenv
pandas
scikit-learn